Move app.py into a module in the husky_musher package
This lets us properly use sibling modules in the same package.  The
root-level app.py still exists as the entrypoint for `flask run` and
WSGI.
